Fix AssetPair::includes_exchange_rates merging sub relations verbatim.
A use case in the code (see AssetPairsController#index) was to pass &:latest as a block, calling the respective scope in ExchangeRate.
The issue with that is that due to the way grouping works, the Left Outer Join property is lost.

Therefore, the method now replaces group column `:asset_pair_id`, if present, by `'asset_pairs.id'`. This still causes the results to be grouped by AssetPair, but keeps the Left Outer Join.